requirements.txt
Flask-HTTPAuth==4.7.0
flask_restful==0.3.9
requests
jsonschema